backend/pyproject.toml
[project]
name = "ijara-backend"
version = "0.1.0"
description = "Ijara rental marketplace — Django 5.2 REST API."
requires-python = ">=3.12,<3.13"
dependencies = [
    "Django>=5.2,<5.3",
    "djangorestframework>=3.16,<3.17",
    "psycopg[binary]>=3.2,<3.3",
    "django-environ>=0.12,<0.13",
    "django-cors-headers>=4.7,<5",
    "django-filter>=25,<26",
    "drf-spectacular>=0.28,<0.29",
    "djangorestframework-simplejwt>=5.5,<5.6",
    "Pillow>=11,<12",
    "python-magic>=0.4,<0.5",
    "stripe>=10,<13",
]

[dependency-groups]
dev = [
    "ruff>=0.9,<1.0",
    "pytest>=8,<9",
    "pytest-django>=4.9,<5",
    "factory-boy>=3,<4",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ijara.settings"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["apps"]
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [
    "**/migrations/**",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B", "UP", "SIM"]
ignore = []

[tool.ruff.lint.per-file-ignores]
# Demo seed data uses long descriptive strings that read better as one line.
"apps/catalog/management/commands/seed_demo.py" = ["E501"]
